reboot with statefn-lex & yacc
Parsing works as intended - extra bits can be added, like
more operators, parens etc.
By the way, it was impossible to have separate intExpr and varExpr
in the grammar, it would be ambiguous on varrefs.

What remains to be done is the evaluator, and putting the final data
to a target struct via reflect.