lex: better naming for accept* methods
What was acceptAny is really acceptRun, as it accepts a run.
The word 'any' was misleading, as it could be accepting any
of ithe conditions met, but just for one run of input.

The difference from Rob Pike's original code is that here I have more
general acceptRun(predicateFunc), while the method accepting runes
from a set given in string is named acceptRunFromSet.
It's cool to give exact names.