embed pos in ast nodes
The goal is to approximate 'relevant' position
in error reports, ofc.

For literals the recorded pos is just after lexing it.
For unary ops, after op and operand.
For binary ops, after 1st operand.
For blocks, after ident defining it's type.
For the top node, pos is zero
(not decided yet if it should be more meaningful).

The seemingly inconsistent positions are actually chosen,
as said above, to mark a relevant place in case eval errors arise.
It should be checked by experimentation if they are best chosen.

Btw, this commit is interesting in a way that it brings back
the node type (interface). All nodes are supposed to know
their position now.