use variable-length ints in the bytecode whenever possible
Ints (actually, uints) are encoded via sqlite4 algo, see
https://www.sqlite.org/src4/doc/trunk/www/varint.wiki .

So far it is used with the opcodes:
- CONST
- DEFBLOCK (for both type & name args)
- GET/SETFIELD
above opcodes use consts, so this is read uniformly in the vm loop; plus also:
- GET/SETLOCAL
- POPN

This removes restriction of locals size <= 256, thus it is rather bound
by the vm stack size (locals live on the stack).

By the way, jumps (both forward and back) can't use variable-length arg,
as the length of the jump value must be known in emitJump, when generating
placeholder that is later patched by patchJump. So for now, jump args
takes 2B, going to 65535 max.