FIX 10c1256 api test generation
There was a subtle bug occuring when the last py test was without
disadm option, then the generated Go loop had a code not considering
the disasm option at all. If was left in a confusing way, disasm
is passed to the Go test table anyway, so no need to differentiate
the function call around this option.