fix leftover bug in lexer.next
Subtle bug, as it is not easy to see token leftovers before switching
to the next input. One such case is in the current big1.bcl, line 53.

Before, that is: with ParseFile reading line by line, this case
was obviously not popping up, and it was hard to find any case for a leftover.
It wasn't manifesting because input boundary was also a token bundary,
with reading by chunks (pages) it's no longer like that, input boundary
can be in the middle of the token.