fix block keys: nameless block is keyed its type
Block without a name now has a key constructed just from its type.
So, no more "blocktype.#<num>" keys.

This fixes reflection issue, where it was tricky to guess a key
that needs to be used as a struct tag.

The consequence is that one scope can't have two nameless blocks
of the same type - but this is sane assumption.