bind: un-arrow (remove syntax "-> slice|struct")
See #6

What still remains to be written is handling of:
  block:first,"name"
because so far a comma-separated list is only allowed for quoted names.

This would require parser to create constants not only for strings
but also for keyword selectors (as 8bit ints for example),
then interpret this accordingly in the vm.